requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.7
django-admin-extra-buttons==1.5.8
django-appconf==1.1.0
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-model-utils==5.0.0
django-qrcode==0.3
django-request==1.7.0
django-select2==8.4.0
django-unfold==0.53.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
